Complete rework of project (#54)
* feat!: complete rework of project

* fix: attributes_values patch

* feat: added attachments functionality and changed the expandable widgets

* feat: can remove comments

* feat: can create comments

* feat: can edit assignee

* feat: can edit watchers and assignees

* fix: remove promote for issue for now

* fix: some cleaning in issues

* fix: streamline WorkItemEditShared

* chore: more modularization

* fix: top app bar

+++ app/src/main/kotlin/com/grappim/taigamobile/data/model/CreateCommentRequest.kt
@@ -3,4 +3,4 @@ package com.grappim.taigamobile.data.model
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class CreateCommentRequest(val comment: String, val version: Int)
+data class CreateCommentRequest(val comment: String, val version: Long)
